allow ridiculously big hash tables

+++ trans_table.c
@@ -5,7 +5,7 @@
 #include "daydreamer.h"
 
 static const int bucket_size = 4;
-static int num_buckets;
+static size_t num_buckets;
 static int generation;
 static const int generation_limit = 8;
 static int age_score_table[8];
